fixed, it now stores items
