+++ src/main/java/com/lothrazar/cyclicmagic/util/UtilItem.java
@@ -5,6 +5,7 @@ import net.minecraft.util.math.MathHelper;
 
 public class UtilItem {
 
+	
 	public static int getMaxDmgFraction(Item tool, int d) {
 
 		return tool.getMaxDamage() - (int) MathHelper.floor_double(tool.getMaxDamage() / d);
